+++ src/models/MetaTag.php
@@ -144,7 +144,13 @@ class MetaTag extends MetaItem
     {
         $shouldRender = parent::prepForRender($data);
         if ($shouldRender) {
-            MetaValueHelper::parseArray($data);
+            // MetaValueHelper::parseArray by default resolves aliases
+            $shouldResolveAliases = true;
+            if (\in_array($data['name'] ?? $data['property'] ?? '', MetaValueHelper::NO_ALIASES, true)) {
+                // Most tags use `name`-property, Facebook uses `property`-property
+                $shouldResolveAliases = false;
+            }
+            MetaValueHelper::parseArray($data, $shouldResolveAliases);
             // Only render if there's more than one attribute
             if (\count($data) > 1) {
                 // devMode